weather-object-detection/requirements.txt
streamlit>=1.31.0
ultralytics>=8.2.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
